hky0312 存档，mlm和mrm加入实体额外掩码，新增wrc任务

+++ model/attention.py
@@ -255,7 +255,7 @@ def multi_head_attention_forward(query,                           # type: Tensor
     attn_output = torch.bmm(attn_output_weights, v)
     assert list(attn_output.size()) == [bsz * num_heads, tgt_len, head_dim]
     attn_output = attn_output.transpose(0, 1).contiguous().view(tgt_len, bsz, embed_dim)
-    attn_output = linear(attn_output, out_proj_weight, out_proj_bias)
+    attn_output = linear(attn_output.float(), out_proj_weight, out_proj_bias)
 
     if need_weights:
         # average attention weights over heads
